+++ examples/portable/device-display/index.js
@@ -1,9 +1,9 @@
-var devices = require('devices');
-var draw = require('./draw');
+var devices = require('mcujs:module').has('devices') ? require('devices') : {};
 
 if (!devices.display) {
   console.log('No configured display is enabled in this firmware.');
 } else {
+  var draw = require('./draw');
   var display = devices.display.open();
   try {
     draw(display);